refactor(gallery): Make gallery structure more modular
