+++ src/app/gallery/types.ts
@@ -1,0 +1,6 @@
+export interface Photo {
+    url: string;
+    orientation: 'portrait' | 'landscape';
+    folder: string;
+    eventDate: { year: number; month: number; day: number };
+}
